requirements.txt
fastapi
uvicorn
twilio
jinja2
python-multipart
python-dotenv